Extracted the PR logic to a separate file
